+++ src/Mailtrap/GlobalUsings.cs
@@ -48,6 +48,8 @@ global using Mailtrap.ContactFields.Requests;
 global using Mailtrap.ContactEvents;
 global using Mailtrap.ContactEvents.Models;
 global using Mailtrap.ContactEvents.Requests;
+global using Mailtrap.EmailLogs;
+global using Mailtrap.EmailLogs.Models;
 global using Mailtrap.Emails;
 global using Mailtrap.Emails.Models;
 global using Mailtrap.Emails.Requests;