+++ tests/Mailtrap.UnitTests/TestConstants/UrlSegmentsTestConstants.cs
@@ -1,4 +1,4 @@
-﻿namespace Mailtrap.UnitTests.TestConstants;
+namespace Mailtrap.UnitTests.TestConstants;
 
 
 internal static class UrlSegmentsTestConstants
@@ -24,4 +24,5 @@ internal static class UrlSegmentsTestConstants
     internal static string EventsSegment { get; } = "events";
     internal static string BatchEmailSegment { get; } = "batch";
     internal static string StatsSegment { get; } = "stats";
+    internal static string EmailLogsSegment { get; } = "email_logs";
 }